Added react router setup, nginx config
