+++ Dockerfile
@@ -1,2 +1,3 @@
 FROM nginx:alpine
+COPY nginx.conf /etc/nginx/conf.d/default.conf
 COPY build/ /usr/share/nginx/html
