+++ nginx.conf
@@ -1,0 +1,27 @@
+server {
+  listen       80;
+  listen  [::]:80;
+  server_name  localhost;
+
+  root   /usr/share/nginx/html;
+  index index.html index.htm;
+
+  location ~* \.(?:manifest|appcache|html?|xml|json)$ {
+    expires -1;
+  }
+
+  location ~* \.(?:css|js)$ {
+    try_files $uri =404;
+    expires 1y;
+    access_log off;
+    add_header Cache-Control "public";
+  }
+
+  location ~ ^.+\..+$ {
+    try_files $uri =404;
+  }
+
+  location / {
+      try_files $uri $uri/ /index.html;
+  }
+}
